1. Add theme changing mechanism. 2. improve code appearance

+++ DSPManager_Lollipop_Marshmallow_Nougat/src/james/dsp/activity/JdspImpResToolbox.java
@@ -1,12 +1,14 @@
 package james.dsp.activity;
 
-public class JdspImpResToolbox {
+public class JdspImpResToolbox
+{
 
-    static {
-		System.loadLibrary("jamesDSPImpulseToolbox");
+    static
+    {
+        System.loadLibrary("jamesDSPImpulseToolbox");
     }
 
     /* Impulse Response Utils */
-    public static native long[] GetLoadImpulseResponseInfo(String path);
+    public static native int[] GetLoadImpulseResponseInfo(String path);
     public static native int[] ReadImpulseResponseToInt();
 }
